HARNESS-21 [2b/4] commit 8: relocate /goldens/* to /goldens/manual/ + new landing page [docs-ok]
URL split per the design discussion (two-card landing):

Routes:
- /goldens                  → NEW landing with two-card layout
                              (Manual / OBD), each linking to its
                              lane page + showing total + reviewed
                              counts.
- /goldens/manual           ← relocated from /goldens (listing)
- /goldens/manual/[id]      ← relocated from /goldens/[id] (detail)
- /goldens/obd              ← NEW (added in commit 9)
- /goldens/obd/[id]         ← NEW (added in commit 9)

Manual listing page changes:
- listGoldens() call now passes lane: "manual"
- Back-link from listing card → detail uses /goldens/manual/{id}
- Detail page's "Back to listing" links updated to point at
  /goldens/manual (was /goldens) so the back nav stays within
  the lane rather than bouncing to the landing.

Landing page (new):
- Two LaneCard components side-by-side on md+ screens, stacked
  on mobile.
- Each card shows total count + reviewed count (across all team
  reviewers, matching the listing's reviewedCount calculation).
- Manual card uses BookOpen icon, OBD card uses Cpu icon.
- Hover lift + shadow for affordance.
- Loading + error states inline.
- All copy keyed off `goldens.landing.*` i18n namespace (added
  in commit 11).

No backend change in this commit; the new landing fetches
listGoldens({lane: "manual"}) and listGoldens({lane: "obd"}) which
are already wired (backend commits 5-6, frontend commit 7).
